requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
dj-email-url==1.0.2
Django==3.2.7
django-cache-url==3.2.3
djangorestframework==3.12.4
environs==8.0.0
marshmallow==3.13.0
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.2
